requirements.txt
# Core Dependencies
streamlit>=1.32.0
openai>=1.14.2
python-dotenv>=1.0.1

# LangChain
langchain>=0.1.13
langchain-community>=0.0.29
langchain-openai>=0.0.8

# Vector Database
pinecone>=4.0.0

# Database
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.23
alembic>=1.13.1

# Document Processing
PyPDF2>=3.0.1
pdf2image>=1.17.0
python-docx>=1.0.1
openpyxl>=3.1.2
tabula-py>=2.9.0
python-multipart>=0.0.6

# Data Processing
pandas>=2.2.1
numpy>=1.26.4

# Visualization
plotly>=5.18.0

# Authentication (optional)
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
